exitoEnConnectionBD
